Merge Release180 to Master

+++ VodovozBusiness/EntityRepositories/Sale/OrderCountResultNode.cs
@@ -3,7 +3,7 @@
 	public class OrderCountResultNode
 	{
 		public int OrderId { get; set; }
-		public int WaterCount { get; set; }
+		public decimal WaterCount { get; set; }
 		public int DistrictId { get; set; }
 	}
 }
